requirements.txt
flask==3.0.0
google-analytics-data==0.18.4
google-auth==2.25.2
google-api-python-client>=2.100.0
gunicorn==21.2.0
anthropic==0.40.0
plotly==5.18.0
kaleido==0.2.1
apscheduler>=3.10.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9.0
